abstract out useUpdatedChildren
